refactor find_app_name_in_path
