+++ PowerToys.sln
@@ -265,8 +265,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "utils", "utils", "{B39DC643
 		src\common\utils\elevation.h = src\common\utils\elevation.h
 		src\common\utils\EventLocker.h = src\common\utils\EventLocker.h
 		src\common\utils\EventWaiter.h = src\common\utils\EventWaiter.h
+		src\common\utils\excluded_apps.h = src\common\utils\excluded_apps.h
 		src\common\utils\exec.h = src\common\utils\exec.h
-		src\common\utils\game_mode.h = src\common\utils\game_mode.h
 		src\common\utils\HDropIterator.h = src\common\utils\HDropIterator.h
 		src\common\utils\HttpClient.h = src\common\utils\HttpClient.h
 		src\common\utils\json.h = src\common\utils\json.h
