+++ src/common/utils/excluded_apps.h
@@ -1,0 +1,30 @@
+#pragma once
+#include <shellapi.h>
+#include <vector>
+#include <string>
+
+inline bool detect_game_mode()
+{
+    QUERY_USER_NOTIFICATION_STATE notification_state;
+    if (SHQueryUserNotificationState(&notification_state) != S_OK)
+    {
+        return false;
+    }
+    return (notification_state == QUNS_RUNNING_D3D_FULL_SCREEN);
+}
+
+// Checks if a process path is included in a list of strings.
+bool find_app_name_in_path(const std::wstring& where, const std::vector<std::wstring>& what)
+{
+    for (const auto& row : what)
+    {
+        const auto pos = where.rfind(row);
+        const auto last_slash = where.rfind('\\');
+        //Check that row occurs in where, and its last occurrence contains in itself the first character after the last backslash.
+        if (pos != std::wstring::npos && pos <= last_slash + 1 && pos + row.length() > last_slash)
+        {
+            return true;
+        }
+    }
+    return false;
+}
